Encrypt JSON to bin

+++ BLE/decrypt_shared.py
@@ -20,21 +20,37 @@ def decrypt_share(enc_file, passcode):
     f.close()
     return decrypted_data
 
+def encrypt_share(text_file, passcode):
+    cryptor = rncryptor.RNCryptor()
+    filename = text_file
+    f = open(filename, mode='rb')
+    data = f.read()
+    encrypted_data = cryptor.encrypt(data, passcode)
+    f.close()
+    return encrypted_data
+
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("-s", "--shared", help="Shared file (no extension) of device details. Default /tmp/share", default="/tmp/shared.bin")
     parser.add_argument("-p", "--passcode", help="Passcode to decrypt shared file. Default "+_default_passcode, default=_default_passcode)
+    parser.add_argument("-e", "--encrypt", help="Encrypt instead of decrypt. Default No", action="store_true", default=False)
     args = parser.parse_args()
 
     sharedbase = args.shared
     passcode = args.passcode
+    encrypt = args.encrypt
     sharedbin = sharedbase+".bin"
     sharedtxt = sharedbase+".json"
-    if os.path.exists(sharedbin):
+    if os.path.exists(sharedbin) and not encrypt:
         decrypted_data = decrypt_share(sharedbin, passcode)
         _devmap = json.loads(decrypted_data)
         with open(sharedtxt, 'w', encoding='utf-8') as f:
             json.dump(_devmap, f, ensure_ascii=False, indent=4)
+    elif os.path.exists(sharedtxt) and encrypt:
+        encrypted_data = encrypt_share(sharedtxt, passcode)
+        with open(sharedbin, 'wb') as f:
+            f.write(encrypted_data)
+            f.close()
     else:
         print("Usage: -s shared")
